ISPN-3268 Move Equivalence API classes to their own package
