+++ core/src/main/java/org/infinispan/commands/write/ReplaceCommand.java
@@ -25,12 +25,12 @@ package org.infinispan.commands.write;
 import org.infinispan.metadata.Metadata;
 import org.infinispan.commands.MetadataAwareCommand;
 import org.infinispan.commands.Visitor;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.configuration.cache.Configuration;
 import org.infinispan.container.entries.MVCCEntry;
 import org.infinispan.context.Flag;
 import org.infinispan.context.InvocationContext;
 import org.infinispan.notifications.cachelistener.CacheNotifier;
-import org.infinispan.util.Equivalence;
 
 import java.util.Set;
