+++ core/src/main/java/org/infinispan/commons/equivalence/AnyEquivalence.java
@@ -21,7 +21,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.infinispan.util;
+package org.infinispan.commons.equivalence;
 
 /**
  * A compare function for objects.
