+++ core/src/main/java/org/infinispan/commons/equivalence/Equivalence.java
@@ -21,7 +21,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.infinispan.util;
+package org.infinispan.commons.equivalence;
 
 import java.io.Serializable;
