+++ core/src/main/java/org/infinispan/commons/equivalence/EquivalentHashMap.java
@@ -21,7 +21,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.infinispan.util;
+package org.infinispan.commons.equivalence;
 
 import java.util.AbstractCollection;
 import java.util.AbstractSet;
@@ -32,10 +32,12 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Set;
 
+import org.infinispan.util.AbstractMap;
+
 /**
  * Custom hash-based map which accepts no null keys nor null values, where
  * equality and hash code calculations are done based on passed
- * {@link org.infinispan.util.Equivalence} function implementations for keys
+ * {@link org.infinispan.commons.equivalence.Equivalence} function implementations for keys
  * and values, as opposed to relying on their own equals/hashCode/toString
  * implementations. This is handy when using key/values whose mentioned
  * methods cannot be overriden, i.e. arrays, and in situations where users
