+++ core/src/main/java/org/infinispan/commons/equivalence/EquivalentHashSet.java
@@ -21,7 +21,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.infinispan.util;
+package org.infinispan.commons.equivalence;
 
 import java.util.AbstractSet;
 import java.util.Iterator;
@@ -29,7 +29,7 @@ import java.util.Iterator;
 /**
  * Custom hash-based set which accepts no null values, where
  * equality and hash code calculations are done based on passed
- * {@link org.infinispan.util.Equivalence} function implementations for values,
+ * {@link org.infinispan.commons.equivalence.Equivalence} function implementations for values,
  * as opposed to relying on their own equals/hashCode/toString implementations.
  * This is handy when using key/values whose mentioned methods cannot be
  * overriden, i.e. arrays, and in situations where users want to avoid using
