+++ core/src/main/java/org/infinispan/container/DefaultDataContainer.java
@@ -25,6 +25,7 @@ package org.infinispan.container;
 import net.jcip.annotations.ThreadSafe;
 import org.infinispan.CacheException;
 import org.infinispan.metadata.Metadata;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.container.entries.InternalCacheEntry;
 import org.infinispan.eviction.ActivationManager;
 import org.infinispan.eviction.EvictionManager;
@@ -36,7 +37,6 @@ import org.infinispan.loaders.CacheLoaderException;
 import org.infinispan.loaders.CacheLoaderManager;
 import org.infinispan.loaders.CacheStore;
 import org.infinispan.util.CollectionFactory;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.Immutables;
 import org.infinispan.util.TimeService;
 import org.infinispan.util.concurrent.BoundedConcurrentHashMap;
