+++ core/src/main/java/org/infinispan/context/AbstractInvocationContextContainer.java
@@ -25,10 +25,10 @@ package org.infinispan.context;
 
 import org.infinispan.commands.FlagAffectedCommand;
 import org.infinispan.commands.VisitableCommand;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.configuration.cache.Configuration;
 import org.infinispan.remoting.transport.Address;
 import org.infinispan.transaction.TransactionTable;
-import org.infinispan.util.Equivalence;
 
 import javax.transaction.TransactionManager;
