+++ core/src/main/java/org/infinispan/context/SingleKeyNonTxInvocationContext.java
@@ -23,9 +23,9 @@
 
 package org.infinispan.context;
 
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.container.entries.CacheEntry;
 import org.infinispan.context.impl.AbstractInvocationContext;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.InfinispanCollections;
 
 import java.util.Collections;
