+++ core/src/main/java/org/infinispan/context/impl/NonTxInvocationContext.java
@@ -22,9 +22,9 @@
  */
 package org.infinispan.context.impl;
 
+import org.infinispan.commons.equivalence.AnyEquivalence;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.container.entries.CacheEntry;
-import org.infinispan.util.AnyEquivalence;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.InfinispanCollections;
 import org.infinispan.util.CollectionFactory;
