+++ core/src/main/java/org/infinispan/transaction/RemoteTransaction.java
@@ -23,12 +23,12 @@
 package org.infinispan.transaction;
 
 import org.infinispan.commands.write.WriteCommand;
+import org.infinispan.commons.equivalence.AnyEquivalence;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.container.entries.CacheEntry;
 import org.infinispan.transaction.xa.GlobalTransaction;
 import org.infinispan.transaction.xa.InvalidTransactionException;
-import org.infinispan.util.AnyEquivalence;
 import org.infinispan.util.CollectionFactory;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.InfinispanCollections;
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
