+++ core/src/main/java/org/infinispan/transaction/synchronization/SyncLocalTransaction.java
@@ -22,9 +22,9 @@
  */
 package org.infinispan.transaction.synchronization;
 
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.transaction.LocalTransaction;
 import org.infinispan.transaction.xa.GlobalTransaction;
-import org.infinispan.util.Equivalence;
 
 import javax.transaction.Transaction;
