+++ core/src/main/java/org/infinispan/transaction/xa/recovery/RecoveryAwareRemoteTransaction.java
@@ -23,10 +23,10 @@
 package org.infinispan.transaction.xa.recovery;
 
 import org.infinispan.commands.write.WriteCommand;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.remoting.transport.Address;
 import org.infinispan.transaction.xa.GlobalTransaction;
 import org.infinispan.transaction.RemoteTransaction;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
