+++ core/src/main/java/org/infinispan/util/concurrent/BoundedConcurrentHashMap.java
@@ -31,8 +31,8 @@
  */
 
 package org.infinispan.util.concurrent;
+import org.infinispan.commons.equivalence.Equivalence;
 import org.infinispan.container.entries.CacheEntry;
-import org.infinispan.util.Equivalence;
 import org.infinispan.util.InfinispanCollections;
 import org.infinispan.util.Util;
 import org.infinispan.util.logging.Log;
