+++ core/src/test/java/org/infinispan/marshall/VersionAwareMarshallerTest.java
@@ -45,6 +45,7 @@ import org.infinispan.commands.write.PutMapCommand;
 import org.infinispan.commands.write.RemoveCommand;
 import org.infinispan.commands.write.ReplaceCommand;
 import org.infinispan.commands.write.WriteCommand;
+import org.infinispan.commons.equivalence.AnyEquivalence;
 import org.infinispan.commons.hash.MurmurHash3;
 import org.infinispan.config.Configuration;
 import org.infinispan.container.entries.ImmortalCacheEntry;
@@ -75,7 +76,6 @@ import org.infinispan.test.fwk.TestCacheManagerFactory;
 import org.infinispan.test.fwk.TestInternalCacheEntryFactory;
 import org.infinispan.transaction.xa.GlobalTransaction;
 import org.infinispan.transaction.xa.TransactionFactory;
-import org.infinispan.util.AnyEquivalence;
 import org.infinispan.util.ByteArrayKey;
 import org.infinispan.util.DefaultTimeService;
 import org.infinispan.util.FastCopyHashMap;
