+++ core/src/test/java/org/infinispan/util/EquivalentHashMapTest.java
@@ -23,6 +23,9 @@
 
 package org.infinispan.util;
 
+import org.infinispan.commons.equivalence.ByteArrayEquivalence;
+import org.infinispan.commons.equivalence.Equivalence;
+import org.infinispan.commons.equivalence.EquivalentHashMap;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
