+++ core/src/test/java/org/infinispan/util/EquivalentHashSetTest.java
@@ -23,6 +23,9 @@
 
 package org.infinispan.util;
 
+import org.infinispan.commons.equivalence.ByteArrayEquivalence;
+import org.infinispan.commons.equivalence.Equivalence;
+import org.infinispan.commons.equivalence.EquivalentHashSet;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
