+++ server/hotrod/src/main/scala/org/infinispan/server/hotrod/HotRodServer.scala
@@ -27,7 +27,8 @@ import scala.collection.JavaConversions._
 import org.infinispan.manager.EmbeddedCacheManager
 import org.infinispan.server.core.AbstractProtocolServer
 import org.infinispan.eviction.EvictionStrategy
-import org.infinispan.util.{CollectionFactory, AnyEquivalence}
+import org.infinispan.util.CollectionFactory
+import org.infinispan.commons.equivalence.AnyEquivalence
 import org.infinispan.Cache
 import org.infinispan.remoting.transport.Address
 import org.infinispan.configuration.cache.{CacheMode, ConfigurationBuilder}
