+++ server/hotrod/src/test/scala/org/infinispan/server/hotrod/HotRodMarshallingTest.scala
@@ -27,7 +27,8 @@ import org.testng.Assert._
 import org.infinispan.commands.remote.ClusteredGetCommand
 import org.infinispan.server.core.AbstractMarshallingTest
 import org.infinispan.api.BasicCacheContainer
-import org.infinispan.util.{ByteArrayEquivalence, InfinispanCollections}
+import org.infinispan.commons.equivalence.ByteArrayEquivalence
+import org.infinispan.util.InfinispanCollections
 
 /**
  * Tests marshalling of Hot Rod classes.
