+++ server/memcached/src/test/scala/org/infinispan/server/memcached/MemcachedReplicationTest.scala
@@ -31,7 +31,7 @@ import net.spy.memcached.CASResponse
 import org.infinispan.manager.EmbeddedCacheManager
 import org.infinispan.test.fwk.TestCacheManagerFactory
 import org.infinispan.configuration.cache.{CacheMode, ConfigurationBuilder}
-import org.infinispan.util.ByteArrayEquivalence
+import org.infinispan.commons.equivalence.ByteArrayEquivalence
 
 /**
  * Tests replicated Infinispan Memcached servers.
